Configuration for Lab 1
I modified the TFT example project. I changed up the clock frequency, so
the system is at 64 MHz and the peripheral bus is 32 MHz. Also added
functions to set the peripheral pins as necessary, set up the timer, set
up the input capture, and the comparator. See Lab1Config.h